开始使用标准库

+++ AD5302.h
@@ -1,8 +1,10 @@
 #ifndef __AD5302_h__
 #define __AD5302_h__
 
+#include "main.h"
+
 void AD5302_Init(void);
 
-void AD5302_Setting(unsigned int set);
+void AD5302_Setting(uint16_t set);
 
 #endif // __SimSPI_h__
