+++ MS5536c.h
@@ -3,8 +3,8 @@
 // #define MS5536c_SCLK
 // #define MS5536c_DOUT
 // #define MS5536c_DIN
-extern unsigned int MS5536c_T;
+extern uint16_t MS5536c_T;
 void MS5536c_Reset(void);
 void MS5536c_GetCx(void);
-unsigned int MS5536c_Calculate(void);
+uint16_t MS5536c_Calculate(void);
 #endif // __MS5536c_h__
